попробовал разделить программу на файлы. есть комментарий в файле, где остались вопросы
