+++ .gitignore
@@ -9,6 +9,7 @@
 *.user
 *.userosscache
 *.sln.docstates
+*copy.*
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
