+++ file_path.py
@@ -1,0 +1,5 @@
+#operation with path and files
+
+source_path = ''
+target_path = ''
+csv_name = 'imgs_numbering.csv'
