+++ work_report.py
@@ -1,0 +1,8 @@
+from counter_mark import csv_crtr, numbering, count
+from image_saver import save_img
+
+def create_file(index,file_name):
+    csv_crtr(index,file_name)
+    save_img(file_name, numbering(index))
+    print(index, ' files ready')
+    count(index)
